requirements.txt
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-environ==0.4.5
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
Markdown==2.6.11
Pillow==5.2.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5
requests==2.19.1
six==1.11.0
static3==0.7.0
Unipath==1.1
urllib3==1.23
webencodings==0.5.1
whitenoise==3.3.1
